Move get global func to the ffi side
